XFF header handler
Play’s remote address method uses a whitelist to determine which IP is
the last proxy server. Without maintaining the Fastly ip list in our
configs, we can’t use this function to get the users ip address.

This is not used in a security context, and is just used for additional
audit information- so getting the first XFF ip is appropriate.