Support Mobile apps authenticating using a SC_GU_U + access-token combo
Clients (ie Mobile apps) can now authenticate with SC_GU_U alone. In order
for test-users to work, you need to send the displayName as well, and this
can come from either the GU_U cookie or a token. The mobile apps will
probably end up sending both SC_GU_U and the access-token, in order to
allow use of Test Users (which require the displayName to determine if
the user is a test user).

This change drops the requirement for the (deprecated) GU_U cookie to
authenticate with cookies - though if the cookie _is_ present, the
displayName will be read from it, as this is still very important for
Membership/Subs test users.